Added UI that allows users to draw new shapes, added methods to draw UIshapes easily

+++ src/main/java/com/projetpaparobin/zones/Point.java
@@ -7,30 +7,30 @@ public class Point {
     
     // Parameters
     
-    private int x,y;
+    private double x,y;
     
     // Constructor
     
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
     
     // Methods
     
-    public void setX(int posX){
+    public void setX(double posX){
         this.x=posX;
     }
     
-    public void setY(int posY){
+    public void setY(double posY){
         this.y=posY;
     }
     
-    public int getX(){
+    public double getX(){
         return this.x;
     }
     
-    public int getY(){
+    public double getY(){
         return this.y;
     }
